+++ Sources/Sentry/include/SentrySDKInternal.h
@@ -42,6 +42,11 @@ SENTRY_NO_INIT
  */
 @property (class, nonatomic, readonly) BOOL isEnabled;
 
+/**
+ * Indicates whether @c start has been called without a matching @c close.
+ */
+@property (class, nonatomic, readonly) BOOL hasStarted;
+
 #if SENTRY_TARGET_REPLAY_SUPPORTED
 /**
  * API to control session replay
@@ -55,6 +60,9 @@ SENTRY_NO_INIT
  *
  * @discussion Call this method on the main thread. When calling it from a background thread, the
  * SDK starts on the main thread async.
+ *
+ * If @c start is called again without @c close in between, the SDK logs a warning and ignores the
+ * call.
  */
 + (void)startWithOptions:(SentryOptionsObjC *)options NS_SWIFT_NAME(start(options:));